Retry a failed target once more in regression tests
We currently have a lot of flaky tests, and I usually find myself
rerunning a failed job with the exactly same code. This happens quite
often and is kind of frustrating. We usually run tests in our local and
we already fix most of the failures in our local, sometimes a test can
indeed fail in circleci but it is not that often.

As a temporary solution to that, this commit reruns the failed target
once more. Circleci currently does not support this feature, so I
thought of adding this to our Makefile.

One downside of this approach is that if a target is indeed failing then
the test will take around 2x longer, however in general this is not the
case and I feel like the benefit of this commit is more than its
downside.