Downgrade from 9.3-1 to 9.2-4 AND upgrade from 9.2-2 to 9.2-4
We have accidentally tagged master branch with v9.2-3. However master branch
was already bumped to v9.3-1, with this commit, we are undoing the catalog
changes that have happened between 9.2-2 to 9.3-1, and making 9.2-4 as the
release that we can use as the continuation of 9.2 series.

Similarly, we have a path to upgrade from 9.2-2 to 9.2-4