fix the refreshing problem with vacuum count
We have a method called wait_for_stats, which checks the current count
and compares it with the previous count, which is a materialized view.
When we perform a vacuum, we call wait_for_stats and this will hopefully
see that the new value is +1 of the previous one that we stored with
materialized view, then we refresh the materialized view so that we will
have the right value when we call the wait_for_stats next time. However
it seems that we were not calling refresh materialized view at some
places after we call wait_for_stats and we can forget to call it. It
makes more sense to embed this into wait_for_stats, so that every time
we call it, we will be refreshing the materialized view as well.