Do not write results locally for router queries
When marked with `SUBPLAN_ACCESS_LOCAL`, all intermediate
results are written to locally, without any further checks.

For example, if the final query hits only local node, that is calculated
during the execution phase, and the intermediate results are multi-casted
accordingly such that the results are written locally.

To recap, this flag was introduced for subqueries in HAVING whose results
are needed on the coordinator for the master plan. So, that is irrelevant
for router queries as router queries cannot have masterPlan.