introduce CitusLocalPlanner to plan queries involving citus local tables
We introduced a new table type to citus, citus local tables. Queries
involving citus local tables cannot be planned with other citus planners
as they even do not know citus tables with distribution method
CITUS_LOCAL_TABLE.

Hence, if a query includes at least one citus local table in it, we first
fall into CreateCitusLocalPlan, and create a distributed plan including
the job to be executed on the coordinator node (note that only the
coordinator is allowed to have citus local tables for now). Then we replace
OID's of citus local tables with their local shards on the query tree and
create the distributed plan with this modified query.

Replacing those tables in the given query, then we create a Job which
executes the given query via executor. Then those queries will be re-
evaluated by the other citus planners without any problems as they know
how to process queries with Postgres local tables.

In that sense, we will behave those tables as local tables accross the
distributed planner and executor. But, for example, we would be erroring
out for their "local shard relations" if it is not a supported query as
we are threating them as Postgres local tables. To prevent this, before
deciding to use CitusLocalPlanner, we first check for unsupported cases
by threating those as local tables and error out if needed.
(see ErrorIfUnsupportedQueryWithCitusLocalTables and its usage)

Reason behind that we do not directly replace the citus local tables and
use existing planner methods is to take necessary locks on shell tables
and keeping citus statistics tracked for citus local tables as well.

ErrorIfUnsupportedQueryWithCitusLocalTables errors out if the given query
is an unsupported "citus local table" query. A query involving citus local
table is unsupported if it is:

 - an UPDATE/DELETE command involving reference tables or distributed
   tables, or
 - an INSERT .. SELECT query on a citus local table which selects from
   reference tables or distributed tables, or
 - a SELECT query involving distributed tables, or
 - a non-simple SELECT query involving reference tables
or:
 - we are not in the coordinator, or
 - coordinator has no placements for citus local tables.