Rebuild wait event sets after PQconnectPoll()
Always rebuild the wait events after PQconnectPoll(). The reason is
that PQconnectPoll() may change the underlying socket, and there is no
proper way for us to understand that.

If we don't rebuild the wait event set, the low level APIs (such as
epoll_ctl()) may fail due to invalid sockets.