Fix for pruned target list entries
The postgres planner has the ability to prune target list entries
that are proven not used in the output relation. When this happens
at the CitusCustomScan boundry we need to _not_ return these pruned
columns to not upset the rest of the planner.

By copying the target list the planner asks us to return we fix
issues that lead to Assertion failures, and potentially could be
runtime errors when they hit in a production build.

Fixes #3809