Add a CI check to see if all tests are part of a schedule
I recently forgot to add tests to a schedule in two of my PRs. One of
these was caught by review, but the other one was not. This adds a
script to causes CI to ensure that each test in the repo is included in
at least one schedule.

Two tests were found that were currently not part of a schedule. This PR
adds those two tests as well.